Update buildx.sh removed included if then

+++ buildx.sh
@@ -18,8 +18,8 @@ echo ${CONTAINER_NAME_TAG_VERSION}
 echo ${BUIILDX_REPO}
 echo ${BUILD_PLATFORM}
 
-# if [ ${CADDY_CURRENT_VERSION} != ${CADDY_VERSION:1} ]
-# then
+if [ ${CADDY_CURRENT_VERSION} != ${CADDY_VERSION:1} ]
+then
   echo "docker buildx build . --platform=${BUILD_PLATFORM} --tag ${CONTAINER_NAME} --tag ${CONTAINER_NAME_TAG_VERSION} --build-arg CADDY_VERSION=${CADDY_VERSION:1} --push"
 
   docker run --rm --privileged multiarch/qemu-user-static --reset -p yes
@@ -33,10 +33,8 @@ echo ${BUILD_PLATFORM}
 
   #remove multiplatform environment
   docker buildx rm ${BUIILDX_REPO}
-
-# else
-  # echo "versions are the same nothing to do here.....!!"
-
-# fi
+else
+  echo "versions are the same nothing to do here.....!!"
+fi
